Explain the 2>> command
Next: the echo token will have to get smarter.
At present, not only are dd and >> allowed (which they should be)
but also d> and >d, which should not be allowed.

Also: the list of commands that abort operator-pending mode needs to be
set dynamically for each operator. The following are all aborted:
dc
cd
d>
><
<>
etc.